Massive UI and functionallity upgrades, as well as performance optimizations and big fixes

+++ app/src/main/java/lilium/arubabacon/DataHandler.java
@@ -3,29 +3,114 @@ package lilium.arubabacon;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteStatement;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.graphics.Matrix;
+import android.util.Log;
+
+import org.apache.commons.io.FileUtils;
+
+import java.io.ByteArrayOutputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Collection;
 
 class DataHandler {
     SQLiteDatabase db;
-    boolean is_open;
+    Boolean is_open;
 
     DataHandler(){
         is_open=false;
 
     }
 
+    public ArrayList<String> availibleDBs(String path){
+        File file = new File(path);
+        ArrayList<String> filenames = new ArrayList<>();
+        for (File f:file.listFiles()){
+            if(f.getName().endsWith(".db")){
+                filenames.add(f.getName().replace(".db",""));
+            }
+        }
+        return filenames;
+    }
+
     public void open(String filename){
         db = SQLiteDatabase.openOrCreateDatabase(filename, null);
         if (db.isOpen())
             is_open = true;
+    }
 
+    public boolean newMap(String filename, File path, File image){
+        close();
+        if (!filename.endsWith(".db"))
+            filename = filename.concat(".db");
+        open(String.format("%s/%s",path.getAbsolutePath(),(filename)));
+        initDB();
+        return insertMap(filename.replace(".db",""), image);
+    }
+
+    public boolean insertMap(String name, File file){
+        if (!is_open) return false;
+        SQLiteStatement insertStatement = db.compileStatement("INSERT INTO map (name, bytes) VALUES (?, ?);");
+        if(file.exists()) {
+            try {
+                ByteArrayOutputStream stream = new ByteArrayOutputStream();
+                Bitmap bitmap = decodeFile(file);
+                if(bitmap.getWidth() > bitmap.getHeight())
+                    bitmap = RotateBitmap(bitmap, 90);
+                bitmap.compress(Bitmap.CompressFormat.PNG,100,stream);
+                byte[] bytes = stream.toByteArray();
+                insertStatement.bindString(1,name);
+                insertStatement.bindBlob(2,bytes);
+                insertStatement.execute();
+            }catch (Exception e){
+                Log.e("DataHandler","Exception at insertMap");
+            }
+        } else {
+            return false;
+        }
+        return true;
+    }
+
+    public byte[] getMap(){
+        if (!is_open) return null;
+        Cursor c = db.query("map",new String[] {"name","bytes"},null,null,null, null, null);
+        if (c.getCount() > 0){
+            c.moveToFirst();
+            byte[] map = c.getBlob(c.getColumnIndex("bytes"));
+            c.close();
+            return map;
+        }
+        c.close();
+        return null;
+    }
+
+    public String getMapName(){
+        if (!is_open) return null;
+        Cursor c = db.query("map",new String[] {"name"},null,null,null, null, null);
+        if (c.getCount() > 0){
+            c.moveToFirst();
+            String name = c.getString(c.getColumnIndex("name"));
+            c.close();
+            return name;
+        }
+        c.close();
+        return null;
     }
 
     public void close(){
-        db.close();
+
+            if (is_open)
+                db.close();
+            is_open = false;
     }
 
     public ArrayList<iBeacon> getBeacons(){
+        if (!is_open) return null;
         Cursor c = db.query("beacons",new String[] {"mac","x","y"},null,null,null, null, null);
         if (c.getCount() > 0){
             ArrayList<iBeacon> beaconList = new ArrayList<iBeacon>();
@@ -41,6 +126,7 @@ class DataHandler {
     }
 
     public iBeacon selectBeacon(String mac, int rssi){
+        if (!is_open) return null;
         Cursor c = db.query("beacons",new String[] {"mac","x","y"},"mac = ?",new String [] {mac},null, null, null);
         if (c.getCount() > 0){
             c.moveToFirst();
@@ -53,16 +139,19 @@ class DataHandler {
     }
 
     public void initDB(){
+        if (!is_open) return;
         db.execSQL("CREATE TABLE IF NOT EXISTS beacons(mac TEXT PRIMARY KEY, x REAL, y REAL);");
         db.execSQL("CREATE TABLE IF NOT EXISTS map (name TEXT, bytes BLOB);");
     }
 
     public void wipeDB(){
+        if (!is_open) return;
         db.execSQL("DROP TABLE IF EXISTS beacons;");
         initDB();
     }
 
     public void addBeacon(String mac, float x, float y){
+        if (!is_open) return;
         SQLiteStatement insertStatement = db.compileStatement("INSERT INTO beacons (mac, x, y) VALUES (?, ?, ?);");
         insertStatement.bindString(1,mac);
         insertStatement.bindDouble(2,x);
@@ -71,10 +160,50 @@ class DataHandler {
     }
 
     public void removeBeacon(iBeacon beacon){
-        SQLiteStatement deleteStatement = db.compileStatement("DELETE b FROM beacons b WHERE b.rssi = ?;");
-        deleteStatement.bindString(1, beacon.mac);
-        deleteStatement.execute();
+        if (!is_open) return;
+        int i = db.delete("beacons","mac = ?",new String[] {beacon.mac});
+        Log.d("dataHandler",String.valueOf(i));
+    }
+
+    public void removeBeacon(String mac){
+        if (!is_open) return;
+        db.delete("beacons","mac = ?",new String[] {mac});
     }
 
 
+    private Bitmap decodeFile(File f){
+        Bitmap b = null;
+
+        //Decode image size
+        BitmapFactory.Options o = new BitmapFactory.Options();
+        o.inJustDecodeBounds = true;
+        try {
+            FileInputStream fis = new FileInputStream(f);
+            BitmapFactory.decodeStream(fis, null, o);
+            fis.close();
+
+            int scale = 1;
+            if (o.outHeight > 2000 || o.outWidth > 2000) {
+                scale = (int) Math.pow(2, (int) Math.ceil(Math.log(2000 /
+                        (double) Math.max(o.outHeight, o.outWidth)) / Math.log(0.5)));
+            }
+
+            //Decode with inSampleSize
+            BitmapFactory.Options o2 = new BitmapFactory.Options();
+            o2.inSampleSize = scale;
+            fis = new FileInputStream(f);
+            b = BitmapFactory.decodeStream(fis, null, o2);
+            fis.close();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return b;
+    }
+
+    public static Bitmap RotateBitmap(Bitmap source, float angle)
+    {
+        Matrix matrix = new Matrix();
+        matrix.postRotate(angle);
+        return Bitmap.createBitmap(source, 0, 0, source.getWidth(), source.getHeight(), matrix, true);
+    }
 }
